Don't make imports at the root of the crate `pub`
Cargo is smart, but not smart enough to check that an item cannot be
used outside of the crate because the crate is a binary. So making an
import `pub` at the root of the crate prevents the lint for unused code
to work.

This then removes everything that was detected as unused.